Separate UI functions into entity modules
