+++ src/ui/cell/factory/createAdjacentMinesElement.ts
