+++ src/ui/renderModalElement.ts
@@ -12,18 +12,19 @@ export const renderModalElement = (gameResult: "win" | "lose"): void => {
   if (gameResult === "lose") {
     modalElement.classList.add("modal--lose");
     modalElement.innerHTML = `
-    <span class="modal-message">Game over!</span>
-    <button class="modal-button" onClick="window.location.reload()">New game</button>
-    `;
+      <span class="modal-message">Game over!</span>
+      <button class="modal-button" onClick="window.location.reload()">New game</button>
+      `;
+
+    mainContentElement.appendChild(modalElement);
+    boardElement.classList.add("disabled");
   }
 
-  if (gameResult === "win") {
-    modalElement.classList.add("modal--win");
-    modalElement.innerHTML = `
+  modalElement.classList.add("modal--win");
+  modalElement.innerHTML = `
     <span class="modal-message">You win!</span>
     <button class="modal-button" onClick="window.location.reload()">New game</button>
     `;
-  }
 
   mainContentElement.appendChild(modalElement);
   boardElement.classList.add("disabled");
